Update Add And Edit Functional

+++ app/Http/Controllers/CartController.php
@@ -13,8 +13,14 @@ class CartController extends Controller
     public function Add(ProductRequest $request){
 
         $product = Product::Where('product_id',$request['product_id'])->first(['user_id','product_id']);
+        $data = Cart::Where('product_id',$request['product_id'])->first();
+        if (isset($data)){
+           Cart::Where('product_id',$request['product_id'])->update(['quantity'=>(int)$data->quantity +1]);
+            return responder()->success();
+        }else{
+            $cart = Cart::create($product->toArray());
+        }
 
-        $cart = Cart::create($product->toArray());
 
         return responder()->success($cart);
 
@@ -22,9 +28,14 @@ class CartController extends Controller
 
     public function Remove(RemoveProductFromCartRequest $request){
 
-        Cart::Where('product_id',$request['product_id'])->delete();
-
+       $data = Cart::Where('product_id',$request['product_id'])->first();
+        if ($data->quantity > 1){
+            Cart::Where('product_id',$request['product_id'])->update(['quantity'=>(int)$data->quantity -1]);
+        }else{
+            Cart::Where('product_id',$request['product_id'])->delete();
+        }
         return response()->noContent();
+
     }
 
     public function Update(UpdateProductInCartRequest $request){
